EE-653: Extract exec_with_return to its own module
