+++ execution-engine/engine-tests/src/support/mod.rs
@@ -1,4 +1,6 @@
 #[cfg(test)]
+pub mod exec_with_return;
+#[cfg(test)]
 pub mod test_stored_contract_support;
 #[cfg(test)]
 pub mod test_support;
